Fixed muzzle flashes was not showing up for GimbalTurret.
git-svn-id: svn://localhost/gltestplus/branches/bullet@1059 42c370b9-4569-d348-85c3-ab3aefc53057

+++ gltestdll/GimbalTurret.h
@@ -59,7 +59,7 @@ protected:
 
 	double yaw;
 	double pitch;
-	float cooldown;
+	double cooldown;
 	float muzzleFlash;
 
 	static Model *model;
